Move commitPostProcessUpdate() into src directory.
