+++ packages/owl-bot/src/bin/commands/commit-post-processor-update.ts
@@ -20,12 +20,8 @@
 // is true and the most recent commit was a copy from googleapis-gen, then
 // we want to squash the changes made by the post processor.
 
-import {cwd} from 'process';
 import yargs = require('yargs');
-import {newCmd} from '../../cmd';
-import {findCopyTag, loadOwlBotYaml, unpackCopyTag} from '../../copy-code';
-import * as proc from 'child_process';
-import path = require('path');
+import { commitAndPushPostProcessorUpdate } from '../../commit-post-processor-update';
 
 export const commitPostProcessorUpdateCommand: yargs.CommandModule<{}, {}> = {
   command: 'commit-post-processor-update',
@@ -35,51 +31,6 @@ export const commitPostProcessorUpdateCommand: yargs.CommandModule<{}, {}> = {
     'depending on the squash flag in .OwlBot.yaml.\n\n' +
     'Run this command in the root directory of a client library repository ' +
     'after running the post processor.',
-  handler: () => commitPostProcessorUpdate(),
-};
+    handler: () => commitAndPushPostProcessorUpdate(),
+  };
 
-export async function commitPostProcessorUpdate(repoDir = ''): Promise<void> {
-  const cmd = newCmd(console);
-  if (['', '.'].includes(repoDir)) {
-    repoDir = cwd();
-  }
-  // Add all pending changes to the commit.
-  cmd('git add -A .', {cwd: repoDir});
-  const status = cmd('git status --porcelain', {cwd: repoDir}).toString(
-    'utf-8'
-  );
-  // `git status` --porcelain returns empty stdout when no changes are pending.
-  if (!status) {
-    return; // No changes made.  Nothing to do.
-  }
-  // Unpack the Copy-Tag.
-  const body = cmd('git log -1 --format=%B', {cwd: repoDir}).toString('utf-8');
-  const copyTagText = findCopyTag(body);
-  if (copyTagText) {
-    try {
-      const copyTag = unpackCopyTag(copyTagText);
-      // Look in .OwlBot.yaml for a squash flag.
-      const yaml = await loadOwlBotYaml(path.join(repoDir, copyTag.p));
-      if (yaml.squash) {
-        // Amend (squash) pending changes into the previous commit.
-        cmd('git commit --amend --no-edit', {cwd: repoDir});
-        // Must force push back to origin.
-        cmd('git push -f', {cwd: repoDir});
-        return;
-      }
-    } catch (e) {
-      console.error(e);
-    }
-  }
-  // Commit new changes as a new commit.
-  // Use spawn() instead of cmd() to avoid the shell potentially
-  // misinterpreting commit message.
-  const commitMessage =
-    '🦉 Updates from OwlBot\n\nSee https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md';
-  console.log(`git commit -m "${commitMessage}"`);
-  proc.spawnSync('git', ['commit', '-m', commitMessage], {cwd: repoDir});
-  // Pull any recent changes to minimize risk of missing refs for the user.
-  cmd('git pull', {cwd: repoDir});
-  // Push changes back to origin.
-  cmd('git push', {cwd: repoDir});
-}
